1.4 Modificacion de valores int a double

+++ src/App.java
@@ -3,13 +3,13 @@ public class App {
 	public static void main(String[] args) {
 		
 	}
-	public static int suma (int a, int b) {
+	public static double suma (double a, double b) {
 		return a+b;
 	}
-	public static int resta (int a, int b) {
+	public static double resta (double a, double b) {
 		return a-b;
 	}
-	public static int multiplicar (int a, int b) {
+	public static double multiplicar (double a, double b) {
 		return a*b;
 	}
 }
